feat: create charts folder and move files
