+++ src/components/Charts/DeviceNetworkHealth.tsx
@@ -1,6 +1,5 @@
 import StandartLineChart from './StandardLineChart'
-
-import IconText from './IconText'
+import IconText from '../IconText'
 import { Paragraph, Separator, XStack, YStack } from 'tamagui'
 import { Shadow as ShadowBox, Text } from '@status-im/components'
 
@@ -13,29 +12,35 @@ type ChartData = {
   id: string
   color: string
   data: DataPoint[]
-  maxValue?: number
 }
 
-type DeviceMemoryHealthProps = {
-  currentMemory: number[]
-  maxMemory: number
+type DeviceNetworkHealthProps = {
+  uploadRate: number[]
+  downloadRate: number[]
 }
-const DeviceMemoryHealth = ({ currentMemory, maxMemory }: DeviceMemoryHealthProps) => {
+const DeviceNetworkHealth = ({ uploadRate, downloadRate }: DeviceNetworkHealthProps) => {
   const chartData: ChartData[] = [
     {
-      id: 'cpu',
+      id: 'uploadRate',
       color: '#8DC6BC',
-      data: currentMemory.map((yValue, index: number) => ({
+      data: uploadRate.map((yValue, index: number) => ({
         x: index + 1,
         y: yValue,
       })),
-      maxValue: maxMemory,
+    },
+    {
+      id: 'downloadRate',
+      color: '#D92344',
+      data: downloadRate.map((yValue, index: number) => ({
+        x: index + 1,
+        y: yValue,
+      })),
     },
   ]
   const currentLoad =
     chartData[0].data.length > 0 ? chartData[0].data[chartData[0].data.length - 1].y : 0
 
-  const message = currentLoad < maxMemory ? 'Good' : 'Poor'
+  const message = currentLoad > 60 ? 'Good' : 'Poor'
 
   return (
     <ShadowBox
@@ -53,7 +58,7 @@ const DeviceMemoryHealth = ({ currentMemory, maxMemory }: DeviceMemoryHealthProp
           justifyContent="space-between"
           style={{
             padding: '8px 16px',
-            position: 'relative',
+            position: 'relative', // Make XStack a positioning context
           }}
         >
           <div style={{ position: 'absolute', top: 0, left: 0, right: 0, bottom: 0 }}>
@@ -61,7 +66,7 @@ const DeviceMemoryHealth = ({ currentMemory, maxMemory }: DeviceMemoryHealthProp
           </div>
           <YStack space={'$3'}>
             <Paragraph color={'#09101C'} size={'$6'} fontWeight={'600'}>
-              Memory
+              Network
             </Paragraph>
             <Paragraph color={'#09101C'} size={'$8'} fontWeight={'700'}>
               {currentLoad} GB
@@ -75,7 +80,7 @@ const DeviceMemoryHealth = ({ currentMemory, maxMemory }: DeviceMemoryHealthProp
           </IconText>
           {message === 'Poor' && (
             <Text size={13} color="#E95460">
-              {((currentLoad / maxMemory || 0) * 100).toFixed(0)}% Utilization
+              {((currentLoad / 60) * 100).toFixed(0)}% Utilization
             </Text>
           )}
         </XStack>
@@ -84,4 +89,4 @@ const DeviceMemoryHealth = ({ currentMemory, maxMemory }: DeviceMemoryHealthProp
   )
 }
 
-export default DeviceMemoryHealth
+export default DeviceNetworkHealth
