+++ src/components/Charts/StandardGauge.stories.tsx
